[REF] html_editor: make code toolbar a subcomponent
This rewrites the code toolbar as a subcomponent of the syntax
highlighting component rather than as an overlay. It simplifies the
logics and allows us to use CSS only to manage its visibility, albeit
with the slight cost of having one code toolbar per code block.

task-5166947

Part-of: odoo/odoo#230799
