+++ addons/html_editor/static/src/others/embedded_components/backend/syntax_highlighting/syntax_highlighting.xml
@@ -2,13 +2,13 @@
 <templates>
     <!-- syntax highlighting template -->
     <t t-name="html_editor.EmbeddedSyntaxHighlighting">
+        <CodeToolbar target="state.host" currentLanguage="embeddedState.languageId"
+            getContent="() => this.textarea.value" onLanguageChange="onLanguageChange.bind(this)"/>
         <pre t-ref="pre"/>
         <textarea t-ref="textarea" class="o_prism_source" contenteditable="true"
             t-on-input="onInput"
             t-on-scroll="onScroll"
             t-on-keydown="onKeydown"
-            t-on-mouseover="onHover" t-on-pointerdown="onHover"
-            t-on-mouseleave="onLeave" t-on-pointerleave="onLeave"
             t-on-focus="props.onTextareaFocus"></textarea>
     </t>
 </templates>
