--- addons/html_editor/static/src/others/embedded_components/plugins/syntax_highlighting_plugin/code_toolbar.scss
@@ -1,19 +1,0 @@
-.o_code_toolbar {
-    position: absolute;
-    top: 0;
-    right: 0;
-    padding: 3px 4px;
-    z-index: 1;
-    display: flex;
-    align-items: center;
-    justify-content: flex-end;
-
-    button {
-        display: inline-flex;
-        align-items: center;
-        height: 20px;
-        padding: 0 3px 0 0;
-        font-size: 12px;
-        --btn-hover-bg: #{$o-gray-200};
-    }
-}
