+++ addons/html_editor/static/src/others/embedded_components/plugins/syntax_highlighting_plugin/syntax_highlighting_plugin.scss
@@ -29,4 +29,32 @@
         resize: none;
         white-space: pre;
     }
+
+    .o_code_toolbar {
+        position: absolute;
+        top: 0;
+        right: 0;
+        padding: 3px 4px;
+        z-index: 1;
+        align-items: center;
+        justify-content: flex-end;
+        display: none;
+
+        button {
+            display: inline-flex;
+            align-items: center;
+            height: 20px;
+            padding: 0 3px 0 0;
+            font-size: 12px;
+            --btn-hover-bg: #{$o-gray-200};
+        }
+    }
+
+    &:hover,
+    &:has(.dropdown-toggle.show),
+    &:focus-within {
+        .o_code_toolbar {
+            display: flex;
+        }
+    }
 }
